redux - clean up of logic
