+++ src/containers/Timer/reducer.js
@@ -12,13 +12,14 @@
 import { combineReducers } from 'redux'
 import { SET_VISIBILITY_FILTER } from './constants'
 
-const visibilityFilter = (state = [], action) => {
-  switch (action.type) {
-    case SET_VISIBILITY_FILTER:
-      return action.filter
-    default:
-      return state
+export default (initialState) => {
+  const visibilityFilter = (state = initialState, action) => {
+    switch (action.type) {
+      case SET_VISIBILITY_FILTER:
+        return action.filter
+      default:
+        return state
+    }
   }
+  return visibilityFilter
 }
-
-export default visibilityFilter
