+++ src/redux/initialState.js
@@ -1,0 +1,15 @@
+import { fromJS } from 'immutable'
+
+const initialState = fromJS({
+  add: {},
+  timer: {},
+  todos: [
+    {
+      id:0,
+      completed: false,
+      text: 'My static todo'
+    }
+  ]
+});
+
+export default initialState
