+++ src/redux/store.js
@@ -3,23 +3,13 @@ import { composeWithDevTools } from 'redux-devtools-extension'
 import thunkMiddleware from 'redux-thunk'
 import { fromJS } from 'immutable'
 
-import rootReducer from '../containers/reducers'
-
-// The initial state of the App
-const initialState = fromJS({
-  todos: [
-    {
-      id:0,
-      completed: false,
-      text: 'My static todo'
-    }
-  ]
-});
+import rootReducer from './reducers'
+import initialState from './initialState'
 
 export const initStore = (state = initialState) => {
   return createStore(
     rootReducer,
-    {},
+    fromJS({add:{},timer:{},todos:{}}),
     composeWithDevTools(applyMiddleware(thunkMiddleware))
   )
 }
